feat: expanded examples to simple & advanced mcp examples
