+++ examples/typescript/servers/mcp/simple.ts
@@ -1,0 +1,138 @@
+/**
+ * MCP Server with x402 Paid Tools - Simple Example
+ *
+ * This example demonstrates the RECOMMENDED way to create an MCP server
+ * using the high-level `createX402MCPServer` factory function.
+ *
+ * Run with: pnpm dev
+ */
+
+import { config } from "dotenv";
+import { SSEServerTransport } from "@modelcontextprotocol/sdk/server/sse.js";
+import { ExactEvmScheme } from "@x402/evm/exact/server";
+import { createX402MCPServer } from "@x402/mcp";
+import express from "express";
+import { z } from "zod";
+
+config();
+
+const evmAddress = process.env.EVM_ADDRESS as `0x${string}`;
+if (!evmAddress) {
+  console.error("❌ EVM_ADDRESS environment variable is required");
+  process.exit(1);
+}
+
+const facilitatorUrl = process.env.FACILITATOR_URL;
+if (!facilitatorUrl) {
+  console.error("❌ FACILITATOR_URL environment variable is required");
+  process.exit(1);
+}
+
+const port = parseInt(process.env.PORT || "4022", 10);
+
+/**
+ * Simulates fetching weather data for a city.
+ *
+ * @param city - The city name to get weather for
+ * @returns Weather data object
+ */
+function getWeatherData(city: string): { city: string; weather: string; temperature: number } {
+  const conditions = ["sunny", "cloudy", "rainy", "snowy", "windy"];
+  const weather = conditions[Math.floor(Math.random() * conditions.length)];
+  const temperature = Math.floor(Math.random() * 40) + 40;
+  return { city, weather, temperature };
+}
+
+/**
+ * Main entry point - Simple API using createX402MCPServer factory.
+ *
+ * @returns Promise that resolves when server is running
+ */
+export async function main(): Promise<void> {
+  console.log("\n📦 Using SIMPLE API (createX402MCPServer factory)\n");
+
+  // ========================================================================
+  // SIMPLE: One-liner setup with factory function
+  // ========================================================================
+  const x402Server = createX402MCPServer({
+    name: "x402 Weather Service",
+    version: "1.0.0",
+    facilitator: facilitatorUrl,
+    schemes: [{ network: "eip155:84532", server: new ExactEvmScheme() }],
+  });
+
+  await x402Server.initialize();
+
+  // Register a paid tool - chainable API
+  x402Server.paidTool(
+    "get_weather",
+    {
+      description: "Get current weather for a city. Requires payment of $0.001.",
+      inputSchema: {
+        city: z.string().describe("The city name to get weather for"),
+      },
+    },
+    {
+      scheme: "exact",
+      network: "eip155:84532",
+      price: "$0.001",
+      payTo: evmAddress,
+      extra: { name: "USDC", version: "2" },
+    },
+    async ({ city }) => ({
+      content: [{ type: "text" as const, text: JSON.stringify(getWeatherData(city as string), null, 2) }],
+    }),
+  );
+
+  // Register a free tool - also through x402Server for consistent API
+  x402Server.tool("ping", "A free tool that returns pong", {}, async () => ({
+    content: [{ type: "text", text: "pong" }],
+  }));
+
+  // Start Express server for SSE transport
+  startExpressServer(x402Server.server, port);
+}
+
+/**
+ * Helper to start Express SSE server
+ *
+ * @param mcpServer - The MCP server instance
+ * @param port - Port to listen on
+ */
+function startExpressServer(mcpServer: ReturnType<typeof createX402MCPServer>["server"], port: number): void {
+  const app = express();
+  const transports = new Map<string, SSEServerTransport>();
+
+  app.get("/sse", async (req, res) => {
+    console.log("📡 New SSE connection");
+    const transport = new SSEServerTransport("/messages", res);
+    const sessionId = crypto.randomUUID();
+    transports.set(sessionId, transport);
+    res.on("close", () => {
+      console.log("📡 SSE connection closed");
+      transports.delete(sessionId);
+    });
+    await mcpServer.connect(transport);
+  });
+
+  app.post("/messages", express.json(), async (req, res) => {
+    const transport = Array.from(transports.values())[0];
+    if (!transport) {
+      res.status(400).json({ error: "No active SSE connection" });
+      return;
+    }
+    await transport.handlePostMessage(req, res, req.body);
+  });
+
+  app.get("/health", (_, res) => {
+    res.json({ status: "ok", tools: ["get_weather (paid)", "ping (free)"] });
+  });
+
+  app.listen(port, () => {
+    console.log(`🚀 x402 MCP Server running on http://localhost:${port}`);
+    console.log(`\n📋 Available tools:`);
+    console.log(`   - get_weather (paid: $0.001)`);
+    console.log(`   - ping (free)`);
+    console.log(`\n🔗 Connect via SSE: http://localhost:${port}/sse\n`);
+  });
+}
